+++ .gitignore
@@ -2,3 +2,4 @@
 *.egg-info
 .eggs
 dist
+venv
