+++ doc/code.rst
@@ -1,0 +1,18 @@
+Documentation for the Code
+**************************
+
+.. automodule:: processrunner
+
+Helper functions
+================
+
+These...
+
+.. autofunction:: processrunner.runCommand
+   :members:
+
+ProcessRunner -- Class
+======================
+
+.. automodule:: processrunner.ProcessRunner
+   :members:
